fix bug with relaunching with changed JT
